ml-deployment/requirements.txt
Flask==2.0.3
numpy==1.19.5
scikit-learn==0.24.2
tensorflow-cpu
